requirements.txt
streamlit==1.40.1
Pillow==11.0.0